protect studies, patients and cras routes. get, update and delete only studies, patients and cras that belong to the user

+++ backend/controllers/patientController.js
@@ -1,11 +1,13 @@
 const asyncHandler = require("express-async-handler");
 const Patient = require("../models/patient.model.js");
+const User = require("../models/user.model.js");
 
 //@desc Get all patients
 //@route GET /api/patients
 //@access Private
 const getPatients = asyncHandler(async (req, res) => {
-  const patient = await Patient.find();
+  const patient = await Patient.find({ user: req.user.id });
+
   res.status(200).json(patient);
 });
 
@@ -21,6 +23,7 @@ const setPatient = asyncHandler(async (req, res) => {
     height: req.body.height,
     blood_group: req.body.blood_group,
     study_name: req.body.study_name,
+    user: req.user.id,
   });
   res.status(200).json(newPatient);
 });
@@ -34,6 +37,21 @@ const updatePatient = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Patient not found");
   }
+
+  const user = await User.findById(req.user.id);
+
+  //  Check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  //Make sure the logged in user matches the patient user
+  if (patientToUpdate.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
   const updatedPatient = await Patient.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -50,6 +68,20 @@ const deletePatient = asyncHandler(async (req, res) => {
   if (!patientToDelete) {
     res.status(400);
     throw new Error("Patient not found");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  //  Check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  // Make sure the logged in user matches the patient user
+  if (patientToDelete.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not found");
   }
 
   await Patient.deleteOne();
